backend/requirements.txt
# Backend requirements
flask
flask-cors
tensorflow
pillow
numpy
